backend/requirements.txt
flask
flask-cors
flask-jwt-extended
psycopg2-binary
werkzeug
python-dotenv
flask-sqlalchemy